requirements.txt
lightning==2.1.2
pandas==2.0.3
PyYAML==6.0.1
torch==2.1.1+cu121
torchmetrics==0.11.4
torchvision==0.16.1
